+++ frontend/src/components/avatar/index.tsx
