+++ frontend/src/pages/appointment/components/skeletons.tsx
@@ -1,8 +1,8 @@
 /**
  * Internal dependencies
  */
-import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Card, CardContent, CardHeader } from "@/components/card";
+import { Skeleton } from "@/components/skeleton";
 
 export function ProfileSkeleton() {
   return (
