+++ frontend/src/pages/group-appointment/index.tsx
@@ -14,8 +14,8 @@ import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
-} from "@/components/ui/tooltip";
-import Typography from "@/components/ui/typography";
+} from "@/components/tooltip";
+import Typography from "@/components/typography";
 import {
   capitalizeWords,
   cn,
@@ -26,13 +26,13 @@ import {
   parseFrappeErrorMsg,
 } from "@/lib/utils";
 import { TimeFormat } from "../appointment/types";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/button";
 import PoweredBy from "@/components/powered-by";
-import { Switch } from "@/components/ui/switch";
+import { Switch } from "@/components/switch";
 import TimeZoneSelect from "../appointment/components/timeZoneSelectmenu";
-import Spinner from "@/components/ui/spinner";
+import Spinner from "@/components/spinner";
 import GroupMeetSkeleton from "./components/groupMeetSkeleton";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Skeleton } from "@/components/skeleton";
 import { getIconForKey, validTitle } from "./utils";
 import { useFrappeGetCall, useFrappePostCall } from "frappe-react-sdk";
 import SuccessAlert from "@/components/success-alert";
